[Warhammer]
** Fixe Save on Zone 161

+++ WarhammerV2/Trunk/WorldServer/World/Objets/Creature.cs
@@ -73,7 +73,7 @@ namespace WorldServer
 
             X = Zone.CalculPin((uint)(Spawn.WorldX), true);
             Y = Zone.CalculPin((uint)(Spawn.WorldY), false);
-            Z = Spawn.WorldZ * 2;
+            Z = (ushort)(Spawn.WorldZ * 2);
 
             Heading = (ushort)Spawn.WorldO;
             SetOffset((ushort)(Spawn.WorldX >> 12), (ushort)(Spawn.WorldY >> 12));
